w3c-reports/12092014/mac-wk2/runned-tests/html/syntax/parsing/template/additions-to-the-in-body-insertion-mode/ignore-frameset-token-expected.txt

CONSOLE MESSAGE: line 115: ReferenceError: Can't find variable: testInIFrame

Harness Error. harness_status.status = 1 , harness_status.message = ReferenceError: Can't find variable: testInIFrame
